+++ fluxo-kmp-conf/src/main/kotlin/fluxo/conf/impl/kotlin/KotlinConstants.kt
@@ -2,9 +2,10 @@ package fluxo.conf.impl.kotlin
 
 internal const val KOTLIN_JVM_PLUGIN_ID = "org.jetbrains.kotlin.jvm"
 internal const val KOTLIN_MPP_PLUGIN_ID = "org.jetbrains.kotlin.multiplatform"
+internal const val KOTLIN_COMPOSE_PLUGIN_ID = "org.jetbrains.kotlin.plugin.compose"
+internal const val KMP_COMPOSE_PLUGIN_ID = "org.jetbrains.compose"
 internal const val KAPT_PLUGIN_ID = "org.jetbrains.kotlin.kapt"
 internal const val KSP_PLUGIN_ID = "com.google.devtools.ksp"
-internal const val JETBRAINS_COMPOSE_PLUGIN_ID = "org.jetbrains.compose"
 internal const val INTELLIJ_PLUGIN_ID = "org.jetbrains.intellij"
 
 internal const val COROUTINES_DEPENDENCY = "org.jetbrains.kotlinx:kotlinx-coroutines-core"