+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Added
 - allow disabling Android Lint when no Android plugin is used.
 - detect when the project is a child of a composite build and has no startup tasks.
+- configure Compose with the new Kotlin compiler plugin.
 
 ### Changed
 - revise the hierarchy of the source sets using both the new `KotlinHierarchyTemplate` and the old way. **Can be a BREAKING CHANGE!**